XmlHttpRequest assíncrono
BUG: Os navegadores "removeram" a opção de XmlHttpRequest síncrono, o que causou instabilidade no app. Problema resolvido. Página de download adicionada. Após realizar o download do app, ele (pelo menos boa parte dos recursos (scripts, css, ...)) carrega a partir do localStorage.

+++ libs/localCache.js
@@ -1,27 +1,35 @@
 const LocalResourceCache =
 {
-    HTTPRequest: function(url)
+    HTTPRequest: function(url, callback)
     {
         var xmlHttp = new XMLHttpRequest();
-        xmlHttp.open( "GET", url, false ); // false for synchronous request
-        xmlHttp.send( null );
-        return xmlHttp.responseText;
+        xmlHttp.onreadystatechange = function() { 
+            if (xmlHttp.readyState == 4 && xmlHttp.status == 200)
+                callback(xmlHttp.responseText);
+        }
+        xmlHttp.open("GET", url, true);
+        xmlHttp.send(null);
 
     },
-    RequestResource: function(resUrl)
+
+    DownloadResource: function(resUrl, callback)
     {
-        if (!resUrl) return null;
-        var resource = localStorage.getItem(resUrl);
-        if (resource && resource.length > 1) // Okay, file size certain will be more than 1 byte 
-            return resource;
-        resource = this.HTTPRequest(resUrl);
-        localStorage.setItem(resUrl, resource);
+        this.HTTPRequest(resUrl, resource => {
+            localStorage.setItem(resUrl, resource);
+            callback(resource);
+        });
+    },
+
+    GetResource: function(url)
+    {
+        var resource = localStorage.getItem(url);
         return resource;
     },
+
     LoadResource: function(resUrl, resourceType)
     {
         if (!resourceType) resourceType = 'script';
-        var resource = this.RequestResource(resUrl);
+        var resource = this.GetResource(resUrl);
         var element = document.createElement(resourceType);
         element.innerHTML = resource;
         var head = document.head;
@@ -45,20 +53,27 @@ const LocalResourceCache =
     }
 }
 
+if (!localStorage.getItem('total_files') && !location.href.endsWith('update.html'))
+{
+    alert('Você precisa atualizar o APP!\nClique OK para atualizar.');
+    localStorage.clear();
+    location.href = 'update.html';
+}
+
 function LoadHtml(url)
 {
     var page = localStorage.getItem(url);
+
     if (!page)
-    {
-        page = LocalResourceCache.RequestResource('raw_pages/'+ url.substring(0, url.length - 5) + '_raw_page.html');
-    }
-    document.write(page);
+        page = LocalResourceCache.GetResource('raw_pages/'+ url.substring(0, url.length - 5) + '_raw_page.html');
+    
+        document.write(page);
 }
 
 function CacheClear(warn = true)
 {
     if (warn)
-    alert('ATENÇÃO!\nLimpar o cache pode resolver alguns problemas, entretanto, irá restaurar as configurações padrão.');
+    if (!confirm('ATENÇÃO!\nLimpar o cache pode resolver alguns problemas, entretanto, irá restaurar as configurações padrão.'))return;
     localStorage.clear();
-    location.reload();
+    location.href = 'update.html';
 }
